# Replace separate index_selects with fused_index_select